(svn r1330) Increase size of some vars from int16 to int32 to guard against future overflows

+++ viewport.h
@@ -25,7 +25,7 @@ void UpdateViewportPosition(Window *w);
 void OffsetGroundSprite(int x, int y);
 
 void DrawGroundSprite(uint32 image);
-void DrawGroundSpriteAt(uint32 image, int16 x, int16 y, byte z);
+void DrawGroundSpriteAt(uint32 image, int32 x, int32 y, byte z);
 void AddSortableSpriteToDraw(uint32 image, int x, int y, int w, int h, byte dz, byte z);
 void *AddStringToDraw(int x, int y, StringID string, uint32 params_1, uint32 params_2, uint32 params_3);
 void AddChildSpriteScreen(uint32 image, int x, int y);
